セレクトボックスを設置するコンポーネントを下層に移動。
